# api/openapi.yml
openapi: 3.0.0
paths:
...
  /accounts:
    get:
      operationId: AccountsController_listAll
      parameters: []
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AccountSummaryDto'
        '403': &ref_3
          description: Caller is not an administrator.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
      security: &ref_4
        - vaultfolio_session: []
      summary: List every account (admin only).
      tags: &ref_5
        - accounts
  /accounts/{id}/role:
    patch:
      operationId: AccountsController_changeRole
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangeRoleRequestDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountSummaryDto'
        '403': *ref_3
        '404':
          description: Account not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountsErrorResponseDto'
        '409':
          description: Would leave no active administrator.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountsErrorResponseDto'
      security: *ref_4
      summary: Change an account's role.
      tags: *ref_5
  /accounts/{id}/domain-scopes:
    patch:
      operationId: AccountsController_changeDomainScopes
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangeDomainScopesRequestDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountSummaryDto'
        '400':
          description: One or more domain ids are not recognized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountsErrorResponseDto'
        '403': *ref_3
        '404':
          description: Account not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountsErrorResponseDto'
      security: *ref_4
      summary: Change an account's domain scopes.
      tags: *ref_5
  /accounts/{id}/archive:
    post:
      operationId: AccountsController_archive
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountSummaryDto'
        '403': *ref_3
        '404':
          description: Account not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountsErrorResponseDto'
        '409':
          description: Would leave no active administrator, or already archived.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountsErrorResponseDto'
      security: *ref_4
      summary: Archive an account.
      tags: *ref_5
  /accounts/{id}/reactivate:
    post:
      operationId: AccountsController_reactivate
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountSummaryDto'
        '403': *ref_3
        '404':
          description: Account not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountsErrorResponseDto'
        '410':
          description: This account's retention window has passed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountsErrorResponseDto'
      security: *ref_4
      summary: Reactivate an archived account, within its retention window.
      tags: *ref_5
  /accounts/{id}:
    delete:
      operationId: AccountsController_deleteSelf
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '204':
          description: Deleted.
        '403': *ref_3
        '409':
          description: Would leave no active administrator.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountsErrorResponseDto'
      security: *ref_4
      summary: Permanently delete an account past its retention window.
      tags: *ref_5
...
  /account-overview/accounts/{id}:
    put:
      operationId: AccountOverviewController_update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAccountOverviewEntryRequestDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountOverviewEntryDto'
        '400':
          description: One or more fields are invalid.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountOverviewValidationErrorResponseDto'
        '404':
          description: This account no longer exists.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountOverviewNotFoundErrorResponseDto'
      security: *ref_9
      summary: >-
        Update an account-overview entry. Every field is optional; an omitted
        field is left unchanged.
      tags: *ref_10
    delete:
      operationId: AccountOverviewController_delete
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '204':
          description: Deleted.
        '404':
          description: This account no longer exists.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountOverviewNotFoundErrorResponseDto'
      security: *ref_9
      summary: Delete an account-overview entry.
      tags: *ref_10
info:
  title: Vaultfolio API
  description: >-
    Generated directly from the backend controllers/DTOs — see
    specs/031-openapi-swagger-integration. "Try it out" reuses the
    vaultfolio_session session cookie from a normal app login; it is not a
    separate credential scheme.
  version: 0.0.9
  contact: {}
...
servers: []
components:
  securitySchemes:
    cookie:
      type: apiKey
      in: cookie
      name: vaultfolio_session
  schemas:
...
    ErrorResponseDetailDto:
      type: object
      properties:
        field:
          type: string
          example: quantity
        message:
          type: string
          example: Must be a positive number.
      required:
        - field
        - message
    ErrorResponseDto:
      type: object
      properties:
        error:
          type: string
          description: Stable machine-readable error code.
          example: not_found
        message:
          type: string
          example: Account not found.
        correlationId:
          type: string
          description: >-
            UUID for end-to-end tracing; matches the X-Correlation-Id response
            header.
          format: uuid
        details:
          description: >-
            Only present for field-level validation errors; non-empty when
            present.
          type: array
          items:
            $ref: '#/components/schemas/ErrorResponseDetailDto'
      required:
        - error
        - message
        - correlationId
    AccountSummaryDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          example: member@example.com
        displayName:
          type: string
        role:
          type: string
          enum:
            - ADMIN
            - MEMBER
        status:
          type: string
          enum:
            - ACTIVE
            - ARCHIVED
        archivedAt:
          type: object
          format: date-time
          nullable: true
        retentionExpiresAt:
          type: object
          format: date-time
          nullable: true
        isLastActiveAdmin:
          type: boolean
        domainScopes:
          description: Domain ids this account is entitled to, independent of `role`.
          type: array
          items:
            type: string
      required:
        - id
        - email
        - displayName
        - role
        - status
        - isLastActiveAdmin
        - domainScopes
    ChangeRoleRequestDto:
      type: object
      properties:
        role:
          type: string
          enum:
            - ADMIN
            - MEMBER
      required:
        - role
    AccountsErrorResponseDto:
      type: object
      properties:
        error:
          type: string
          enum:
            - not_found
            - last_admin
            - already_archived
            - retention_expired
            - forbidden
            - invalid_domain
        message:
          type: string
      required:
        - error
        - message
    ChangeDomainScopesRequestDto:
      type: object
      properties:
        domainScopes:
          type: array
          items:
            type: string
      required:
        - domainScopes
...
    AccountOverviewEntryDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        category:
          type: string
          enum: &ref_17
            - GENERAL
            - LEISURE
            - SAVINGS
            - DEPOT
            - CREDIT_CARD
            - OTHER
        status:
          type: string
          enum: &ref_18
            - ACTIVE
            - DECOMMISSIONED
          description: >-
            A decommissioned account stays listed, sorted after the active ones
            within its category.
        provider:
          type: object
          nullable: true
        website:
          type: object
          nullable: true
        purpose:
          type: object
          nullable: true
        cardUsage:
          type: object
          nullable: true
        requiredMinimum:
          type: object
          nullable: true
        notes:
          type: object
          nullable: true
        cardNumber:
          type: object
          nullable: true
          description: Full card number as entered (CREDIT_CARD accounts only).
        validUntil:
          type: object
          nullable: true
          example: 12/29
          description: MM/YY.
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - category
        - status
        - createdAt
        - updatedAt
...
    AccountOverviewValidationErrorResponseDto:
      type: object
      properties:
        error:
          type: string
          enum:
            - VALIDATION_FAILED
        message:
          type: string
        fieldErrors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
      required:
        - error
        - message
        - fieldErrors
    UpdateAccountOverviewEntryRequestDto:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
          enum: *ref_17
          description: Omit to default to 'OTHER'.
        status:
          type: string
          enum: *ref_18
          description: Omit to default to 'ACTIVE'.
        provider:
          type: string
        website:
          type: string
        purpose:
          type: string
        cardUsage:
          type: string
        requiredMinimum:
          type: string
        notes:
          type: string
        cardNumber:
          type: string
        validUntil:
          type: string
          example: 12/29
      required:
        - name
    AccountOverviewNotFoundErrorResponseDto:
      type: object
      properties:
        error:
          type: string
          enum:
            - ACCOUNT_NOT_FOUND
        message:
          type: string
      required:
        - error
        - message